データ型修正

+++ database/migrations/2025_05_28_104441_create_shops_table.php
@@ -13,14 +13,14 @@ return new class extends Migration
     {
         Schema::create('shops', function (Blueprint $table) {
             $table->id();
-            $table->bigint('owner_id');
+            $table->BigInteger('owner_id');
             $table->string('name');
             $table->text('information');
             $table->string('filename');
             $table->string('active');
             $table->timestamps();
-            $table->int('category_id');
-            $table->int('area_id');
+            $table->integer('category_id');
+            $table->integer('area_id');
             $table->string('local');
             $table->boolean('on_off');
