+++ database/migrations/2025_05_28_104525_create_reviews_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->id();
-            $table->bigint('shop_id');
+            $table->BigInteger('shop_id');
             $table->string('reviewer_name');
             $table->double('review');
             $table->text('comment');
